requirements.txt
streamlit
google-generativeai
gTTS
PyPDF2